+++ src/components/Heading/HeadingTypes.tsx
@@ -6,10 +6,10 @@ export enum HeadingDisplaySizes {
 }
 
 export enum HeadingLevels {
-  One = 1,
-  Two = 2,
-  Three = 3,
-  Four = 4,
-  Five = 5,
-  Six = 6,
+  One = "one",
+  Two = "two",
+  Three = "three",
+  Four = "four",
+  Five = "five",
+  Six = "six",
 }