+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Fixes
 
 - Fixes the styling of custom anchor elements when passed as a child in the `Link` component.
+- Fixes the Storybook `level` prop value for the `Heading` component so it updates the component.
 
 ## 0.25.8 (January 6, 2022)